+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2018-07-17
+
+### Fixed
+
+- Give proper error message when the `--number` argument is smaller than one. Fixes #164.
+
 ## [1.5.3] - 2018-07-14
 
 ### Fixed